event-validation-app/requirements/base.txt
# Django specific
django==5.0.8  # https://www.djangoproject.com/
django-environ==0.11.2  # https://github.com/joke2k/django-environ
django-cors-headers==4.4.0  # https://github.com/adamchainz/django-cors-headers
django-filter==24.3  # https://django-filter.readthedocs.io/en/stable/guide/install.html
django-redis==5.4.0

# DRF specific
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1  # https://django-rest-framework-simplejwt.readthedocs.io/en/latest/
drf-spectacular==0.27.2  # https://github.com/tfranzel/drf-spectacular

# Database
psycopg2-binary==2.9.9  # https://github.com/psycopg/psycopg

# Celery specific
celery==5.4.0  # pyup: < 6.0  # https://github.com/celery/celery
django-celery-beat==2.6.0  # https://github.com/celery/django-celery-beat

#Utilities
invoke==2.2.0
python-dotenv==1.0.1
requests==2.32.3
